[FIX] Corrected bad words issue
